udfpshandler: Add authentication succeeded/failed notifications
We currently miss a call when the fingerprint gets rejected
the fifth time, and the onFingerUp method isn't called anymore.
Without handling this in the touchscreen driver, we can not
disable lhbm (ghbm should turn off fine since the udfps layer
vanishes). The onAuthentication* methods are called when the
authentication ends, allowing for a better control of the lhbm
state.

Change-Id: I5ff6b9b4692657fade4579faff33eb48c1311987